+++ R/plotPairs.R
@@ -44,6 +44,7 @@
 #' @importFrom ggplot2 ggplot annotate theme_void ylim stat_density2d
 #'     scale_fill_continuous geom_point theme_bw theme element_blank aes
 #'     geom_histogram scale_x_continuous scale_y_continuous geom_abline
+#'     after_stat
 #' @importFrom stats cor
 #' @importFrom SummarizedExperiment assayNames assay
 #' @importFrom grDevices hcl.colors rgb colorRamp
@@ -102,7 +103,8 @@ plotPairs <- function(se, selAssay = "counts", doLog = TRUE, pseudocount = 1,
         yData <- GGally::eval_data_col(data, mapping$y)
 
         ## Calculate correlation
-        mainCor <- stats::cor(xData, yData, method = corMethod)
+        mainCor <- stats::cor(xData, yData, method = corMethod, 
+                              use = "pairwise.complete.obs")
         transfCor <- (abs(mainCor) - min(corRange))/(max(corRange) - min(corRange))
         ## Protect against numerical imprecision leading to values outside
         ## the allowed range
@@ -141,7 +143,7 @@ plotPairs <- function(se, selAssay = "counts", doLog = TRUE, pseudocount = 1,
     ## Define function to create smoothscatter-like plot (for use with ggpairs)
     smoothscat <- function(data, mapping, ...) {
         ggplot2::ggplot(data = data, mapping = mapping) +
-            ggplot2::stat_density2d(ggplot2::aes(fill = (.data$..density..)^0.25), geom = "tile",
+            ggplot2::stat_density2d(ggplot2::aes(fill = ggplot2::after_stat(.data$density)^0.25), geom = "tile",
                                     contour = FALSE, n = 200) +
             ggplot2::scale_fill_continuous(low = "white", high = "darkgreen") +
             ggtheme +
@@ -204,7 +206,7 @@ plotPairs <- function(se, selAssay = "counts", doLog = TRUE, pseudocount = 1,
                     substr(corMethod, 2, nchar(corMethod)), " correlation")
 
     ## Calculate correlations and get range
-    allCors <- stats::cor(mat, method = corMethod)
+    allCors <- stats::cor(mat, method = corMethod, use = "pairwise.complete.obs")
     if (!is.null(corrColorRange)) {
         corRange <- corrColorRange
     } else {